Crate noirc_artifacts

This module defines the structure of Nargo’s different compilation artifacts.

These artifacts are intended to remain independent of any applications being built on top of Noir. Should any projects require/desire a different artifact format, it’s expected that they will write a transformer to generate them using these artifacts as a starting point.

Modules

contract

debug

debug_vars 🔒

program

ssa

Functions

deserialize_hash 🔒

Deserialize hash from String in JSON.

serialize_hash 🔒

Serialize hash as String, so that it doesn’t get truncated in Javascript.