noirc_artifacts

Module debug

Structs

DebugArtifact

A Debug Artifact stores, for a given program, the debug info for every function along with a map of file Id to the source code so locations in debug info can be mapped to source code they point to.

DebugFile

For a given file, we store the source code and the path to the file so consumers of the debug artifact can reconstruct the original source code structure.

DebugFnId

DebugFunction

DebugInfo

DebugTypeId

DebugVarId

DebugVariable

DebugVars

LocationNodeDebugInfo

LocationTree

ProcedureDebugId

ProgramDebugInfo

StackFrame

Type Aliases

DebugFunctions

DebugTypes

DebugVariables