noirc_artifacts::debug
Struct DebugArtifact

pub struct DebugArtifact {
    pub debug_symbols: Vec<DebugInfo>,
    pub file_map: BTreeMap<FileId, DebugFile>,
}

A Debug Artifact stores, for a given program, the debug info for every function along with a map of file Id to the source code so locations in debug info can be mapped to source code they point to.

Fields
debug_symbols: Vec<DebugInfo>file_map: BTreeMap<FileId, DebugFile>
Implementations

impl DebugArtifact

pub fn new(debug_symbols: Vec<DebugInfo>, file_manager: &FileManager) -> Self

pub fn location_source_code(&self, location: Location) -> Result<&str, Error>

Given a location, returns its file’s source code

pub fn location_line_index(&self, location: Location) -> Result<usize, Error>

Given a location, returns the index of the line it starts at

pub fn location_end_line_index( &self, location: Location, ) -> Result<usize, Error>

Given a location, returns the index of the line it ends at

pub fn location_line_number(&self, location: Location) -> Result<usize, Error>

Given a location, returns the line number it starts at

pub fn location_column_number(&self, location: Location) -> Result<usize, Error>

Given a location, returns the column number it starts at

pub fn location_in_line( &self, location: Location, ) -> Result<Range<usize>, Error>

Given a location, returns a Span relative to its line’s position in the file. This is useful when processing a file’s contents on a per-line-basis.

pub fn location_in_end_line( &self, location: Location, ) -> Result<Range<usize>, Error>

Given a location, returns a Span relative to its last line’s position in the file. This is useful when processing a file’s contents on a per-line-basis.

pub fn last_line_index(&self, location: Location) -> Result<usize, Error>

Given a location, returns the last line index of its file